Restructure code in src/bygg/cmd
This makes dispatcher.py a bit smaller and makes the code easier to work
with.
